refactor: add object function
